+++ packages/nextclade/src/analyze/aa_changes_group.rs
@@ -1,0 +1,49 @@
+use crate::analyze::aa_change_with_context::AaChangeWithContext;
+use crate::analyze::nuc_del::NucDelRange;
+use crate::analyze::nuc_sub::NucSub;
+use crate::coord::range::AaRefRange;
+use itertools::{Itertools, MinMaxResult};
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize, schemars::JsonSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct AaChangesGroup {
+  pub name: String,
+  pub range: AaRefRange,
+  pub changes: Vec<AaChangeWithContext>,
+  pub nuc_subs: Vec<NucSub>,
+  pub nuc_dels: Vec<NucDelRange>,
+}
+
+impl AaChangesGroup {
+  pub fn new(name: impl AsRef<str>) -> Self {
+    Self::with_changes(name, vec![])
+  }
+
+  pub fn with_changes(name: impl AsRef<str>, changes: Vec<AaChangeWithContext>) -> Self {
+    Self {
+      name: name.as_ref().to_owned(),
+      range: Self::find_codon_range(&changes),
+      changes,
+      nuc_subs: vec![],
+      nuc_dels: vec![],
+    }
+  }
+
+  pub fn push(&mut self, change: AaChangeWithContext) {
+    self.changes.push(change);
+    self.range = Self::find_codon_range(&self.changes);
+  }
+
+  pub fn last(&self) -> Option<&AaChangeWithContext> {
+    self.changes.last()
+  }
+
+  fn find_codon_range(changes: &[AaChangeWithContext]) -> AaRefRange {
+    match changes.iter().minmax_by_key(|change| change.pos) {
+      MinMaxResult::NoElements => AaRefRange::from_isize(0, 0),
+      MinMaxResult::OneElement(one) => AaRefRange::new(one.pos, one.pos + 1),
+      MinMaxResult::MinMax(first, last) => AaRefRange::new(first.pos, last.pos + 1),
+    }
+  }
+}
