+++ packages/nextclade/src/analyze/mod.rs
@@ -1,4 +1,7 @@
-pub mod aa_changes;
+pub mod aa_change_with_context;
+pub mod aa_changes_find;
+pub mod aa_changes_find_for_cds;
+pub mod aa_changes_group;
 pub mod aa_del;
 pub mod aa_sub;
 pub mod abstract_mutation;
